gastracker-ui/backend/requirements.txt
fastapi
uvicorn[standard]
langchain-mcp-adapters
langchain-openai
python-dotenv
